requirements.txt
pandas==1.4.2
transformers==4.18.0
scikit-learn==1.1.0
torch==1.11.0+cu113
numpy==1.23.4
onnxruntime-gpu==1.12.1
streamlit==1.13.0